Python/requirements.txt
flask==2.0.1
requests==2.26.0
streamlit
openai
langchain
langchain-core
langchain-community
langchain-openai
langchain-pinecone
langchain-text-splitters
langchain-chroma
pinecone-notebooks
python-dotenv
oracledb
fastapi
uvicorn
pydantic

